Restructure to not need ES module design
