+++ src/server.js
@@ -1,4 +1,4 @@
-import app from "./app.cjs";
+const app = require('./app');
 
 app.listen(8080, () => {
     console.log('\x1b[36m%s\x1b[0m', `
